python/requirements.txt
pandas==2.1.0
psycopg2-binary==2.9.7
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.26.0
plotly==5.15.0
sqlalchemy==2.0.20
